sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "boxed-sdk"
version = "0.1.0"
description = "Python SDK for Boxed Sovereign Execution Engine"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.31.0",
    "websocket-client>=1.6.0",
]
license = {text = "MIT"}

[project.urls]
Homepage = "https://github.com/akshayaggarwal99/boxed"
